Modularize property tests
